split Designer components into seperate files
